Update patch set 1
Patch Set 1:

the more i think about the idea  of forking or creating a branch for this the more i like it ☺ would avoid any upstream merging issues and devices that need it would just need to add the repo or branch to dependencies. the target names are anyway different so there should be no name clashes also.

Patch-set: 1